simplify header into message
